config.py
import os
from dotenv import load_dotenv

# Load file .env 
load_dotenv()

# import variable
TELEGRAM_TOKEN = os.getenv("TELEGRAM_TOKEN")
CHAT_ID = os.getenv("CHAT_ID")
NVD_API = os.getenv("NVD_API")

# URL for API Public
NVD_API_URL = "https://services.nvd.nist.gov/rest/json/cves/2.0"
CISA_KEV_URL = "https://www.cisa.gov/sites/default/files/feeds/known_exploited_vulnerabilities.json"
