requirements.txt
flask
requests
beautifulsoup4
feedparser
python-telegram-bot
python-dotenv
gunicorn